requirements.txt
beautifulsoup4>=4.13.3
certifi>=2025.1.31
charset-normalizer>=3.4.1
contourpy>=1.3.1
cycler>=0.12.1
filelock>=3.17.0
fonttools>=4.57.0
frozendict>=2.4.6
fsspec>=2025.3.0
huggingface-hub>=0.29.3
idna>=3.10
Jinja2>=3.1.6
joblib>=1.4.2
kiwisolver>=1.4.8
MarkupSafe>=3.0.2
matplotlib>=3.10.1
mpmath>=1.3.0
multitasking>=0.0.11
mysql-connector-python>=9.2.0
networkx>=3.4.2
numpy>=2.1.0
packaging>=24.2
pandas>=2.2.3
peewee>=3.17.9
pillow>=11.1.0
platformdirs>=4.3.6
PyMySQL>=1.1.1
pyparsing>=3.2.3
python-dateutil>=2.9.0.post0
python-dotenv>=1.0.1
pytz>=2025.1
PyYAML>=6.0.2
regex>=2024.11.6
requests>=2.32.3
safetensors>=0.5.3
scikit-learn>=1.6.1
scipy>=1.15.2
six>=1.17.0
soupsieve>=2.6
sympy>=1.13.1
threadpoolctl>=3.5.0
tokenizers>=0.21.0
torch>=2.6.0
tqdm>=4.67.1
transformers>=4.49.0
typing_extensions>=4.12.2
tzdata>=2025.1
urllib3>=2.3.0
yfinance>=0.2.61
boto3>=1.35.0
botocore>=1.35.0
rich>=13.9.4
